[wip/libjointstatemapper] Release of libjointstatemapper-1.0
libJointStateMapper is a library that maps a joint configuration from one joint
convention to another one.

It relies at initialization on two lists of joint name (one for the source, one
for the target) and a mapping between these two.

Then, it can be invoked to 'reorder' a joint state (ori::JointState) from the
source convention to the destination convention.