[wip/py-poco] Fix depend.mk
Fix DEPEND_DIR.py-poco, causing errors in show-depends of wip/hri-simulation,
and fix a typo in the pkgname (s/pypoco/py-poco/).

While here, generate headings, disable configure step, and print-PLIST.

I'd suggest that the package is renamed to py-tclserv, as this is what it
really does (at least it does not do any pocolibs wrapping as the name
suggests). And this would be consistent with COMMENT.