[wip/ros18-base] Fix installation and PLIST of some python packages
Some ros python packages require py-setuptools instead of using vanilla
distutils. This has the side effect of installing so called EGG_INFO
directories that behave differently depending on the platform (sometimes the
directory is zipped, sometimes not). The PLIST may then be broken.

Instead of fighting with this, just use distutils instead of setuptools since
setuptools does not bring anything useful in robotpkg (does it?). This is done
via a SUBST stage in setup.py for packages that declare using it.

Affected packages are, so far:
 - ros-gencpp
 - ros-genlisp
 - ros-genpy