[wip/ros18-base] Define PYTHON_EXECUTABLE et al. in Makefile.common
Make sure to pass all variables defining the proper python version as found by
robotpkg, otherwise ros CMakeFiles may happily find python-3.2 and use it
instead of python<3, although this does not work in the end (at least ros-rx
does this and breaks on F17).

This is done in the common ros Makefile as this affects all packages.