[planning/hatpconsole] Add missing dependency
hatpconsole was missing libgvc that can be found in graphviz.
Modify the Makefile to add the dependency, and add the file to
find the libgvc (from graphviz). If a project needs libgvc it
should use the file mk/sysdep/libgraphviz.mk instead of the
mk/sysdep/graphviz.mk in order to get the libraries.
As this was a minor change, update the PKGREVISION.