[wip] Add some sysdep for libgqlviewer
From the website:

libQGLViewer is a C++ library based on Qt that eases the creation of
OpenGL 3D viewers.

It provides some of the typical 3D viewer functionalities, such as the
possibility to move the camera using the mouse, which lacks in most of
the other APIs. Other features include mouse manipulated frames,
interpolated keyFrames, object selection, stereo display, screenshot
saving and much more. It can be used by OpenGL beginners as well as to
create complex applications, being fully customizable and easy to
extend.