[wip/jrl-dynamics-urdl] Make this compile and polish the package
Add two local patches to make this compile. The purpose is to remove the use of
rosbuild in the makefiles. The strategy here is to use plain cmake. However,
catkin is also an option. In any case, this should be merged somehow upstream
before inclusion in robotpkg.

Also added a PLIST, fixed the depend.mk and added a few more dependencies that
are now handled by robotpkg (ros-robot-model and bullet).

Also note that the package compiles fine but was not tested at runtime. In
particular, the test target is missing some data file.