[wip/ros-realtime-tools] Relax ros dependency to ros>=groovy
The package compiles with ros==groovy. However, ros-groovy does not contain
this package, and the ros-realtime-tools dependency will not be met when
PREFER.ros==system.

It is also not currently possible to set a specific PREFER.ros-realtime-tools
default to robotpkg or system automatically depending on the ros version to be
used. So binary packages depending on this one won't built because of the unmet
dependency.

While here, fix a typo in PREFER.ros-realtime-tools and fix files permissions.