[wip] Add a package for lua-rfsm
rFSM is a small and powerful Statechart implementation. It is mainly
designed for Coordination of complex systems but is not limited to that.
rFSM is written in pure Lua and is therefore highly portable and
embeddable. As a Lua domain specific language rFSM inherits the
extensibility of its host language. In particular, it is used within the
Orocos framework for coordination purporse.