[wip/genom3] Improve boost thread configuration for the ros/server template
Pass all the arguments to configure that ax_boost_thread expects, so that the
automatic detection of boost_thread is disabled. This this not only
cleaner/more robust but also fixes the configuration of genom3 components on
ubuntu-1310, where libboost_* have been moved to a subdirectory of /usr/lib
and the ax_boost_thread macros fail to find it automatically.