[math/jrl-dynamics-urdf] Release 2.2.0
v2.2.0

    Merge pull request #13 from francois-keith/master
    Correct the construction of the actuated joints list.
    Merge pull request #11 from francois-keith/master
    Correct package.xml (REP127)
    Correct the construction of the hand frames.
    Correct handling of continuous joint.
    Add soversion.
    Warning removal.
    Correct warning message when the inertial info are missing.
    Correct computation of the hand axes.
    Correct inertia and com frames.
    Display inertia too.
    Add const in display functions.

NB: compared to the 2.2.0 git release, a patch has been added in robotpkg in order to not force BOOST_FILESYSTEM_VERSION=2, since this is not supported in boost>=1.50. It has not been done in the git release because SOT main developpers are not sure that won't produce numerical complications.